--- apps/muse-mobile/src/App.tsx
@@ -1,6 +1,0 @@
-import App from './main';
-import { setupSentry } from './integrations/sentry';
-
-const AppWithSentry = setupSentry(App);
-
-export default AppWithSentry;
