+++ apps/muse-mobile/src/integrations/navigation.tsx
@@ -1,0 +1,15 @@
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import { Home } from '../pages/home';
+
+const Stack = createNativeStackNavigator();
+
+export const Navigation = () => {
+  return (
+    <NavigationContainer>
+      <Stack.Navigator screenOptions={{}} initialRouteName={'Home'}>
+        <Stack.Screen name={'Home'} component={Home} />
+      </Stack.Navigator>
+    </NavigationContainer>
+  );
+};
